azure-functions/requirements.txt
azure-functions==1.17.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-storage-blob==12.19.1
msrest==0.7.1
boto3==1.34.69
requests==2.31.0
